Replace CURL with Aria2
I had to add `mkdir -p` in the download command because, as far as I know, aria2 won't create the directory structure like curl does.

I decided to split the download in the same number of parts as the THREADS env var. It might be that we want to do that separately.